Refactor Panel class and move finding methods into new class
